テーブルコンポーネント修正
