+++ components/pages/home/food/register/FoodRegisterDisplay.vue
@@ -275,14 +275,12 @@ export default {
       this.isClickBtn = false
       // ユーザー認証が切れていたらsigninに遷移
       if (res.error) {
-        alert(res.error)
         this.$router.push({ path: '/signin' })
         return
       }
       console.log(res.result)
       // 登録成功で表示ページへ遷移
       if (res.result) {
-        alert(res.message)
         const food = res.result
         this.$router.push({ path: `/home/food/${food.id}` })
       }
