+++ components/pages/home/recipe/register/RecipeRegisterDisplay.vue
@@ -449,13 +449,11 @@ export default {
       this.isClickBtn = false
       // ユーザー認証が切れていたらsigninに遷移
       if (res.error) {
-        alert(res.error)
         this.$router.push({ path: '/signin' })
         return
       }
       // 成功すれば画面遷移
       if (res.result) {
-        alert(res.message)
         const recipe = res.result
         this.$router.push({ path: `/home/recipe/${recipe.id}` })
       }
