+++ components/pages/home/recipe/update/RecipeUpdateDisplay.vue
@@ -480,7 +480,6 @@ export default {
       this.isClickBtn = false
       // ユーザー認証が切れていたらsigninに遷移
       if (res.error) {
-        alert(res.error)
         this.$router.push({ path: '/signin' })
         return
       }
@@ -499,13 +498,11 @@ export default {
       console.log(res)
       // ユーザー認証が切れていたらsigninに遷移
       if (res.error) {
-        alert(res.error)
         this.$router.push({ path: '/signin' })
         return
       }
       // 成功すれば画面遷移
       if (res.result) {
-        alert(res.message)
         this.$router.push({ path: '/home/recipe' })
       }
     },
